% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{confounderData}
\alias{confounderData}
\title{confounderData}
\format{
An object of class \code{data.frame} with 50 rows and 5 columns.
}
\usage{
confounderData
}
\description{
confounderData
}
\examples{
  head(confounderData)
}
\keyword{datasets}
