% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupInfo}
\alias{groupInfo}
\title{groupInfo}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
groupInfo
}
\description{
groupInfo
}
\examples{
  head(groupInfo)
}
\keyword{datasets}
