% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIS.R
\name{IIS}
\alias{IIS}
\title{Title IIS}
\usage{
IIS(
  Micro_path,
  Meta_path,
  confounderData = NULL,
  groupInfo = NULL,
  nDFS = 0.4,
  nDS = 0.3,
  nES = 0.2,
  nAS = 0.1,
  NS = 2.5,
  r = 0.5,
  p_adjust = 0.05,
  MMpath
)
}
\arguments{
\item{Micro_path}{Microbial data in data.frame format such as 16S or Metagenome data including microorganism abundance and functional annotation}

\item{Meta_path}{Metabolic data in data.frame format that includes Metabolite abundance and related pathways that was conversed using Meta2pathway function}

\item{confounderData}{information of partial correlation in data.frame format}

\item{groupInfo}{the data.frame including two columns. The name of first colunm is 'SampleID' meaning the name of samples in Micro_path or Meta_path, and the name of second colunm is 'Group' meaning the groups the samples belong to.}

\item{nDFS}{the weight value indicated the weight of diff Score in the IIS formula.Default value is 0.4}

\item{nDS}{the weight value  indicated the weight of degree score  in the IIS formula.Default value is 0.3}

\item{nES}{the weight value  indicated the weight of edge score in the IIS formula.Default value is 0.2}

\item{nAS}{the weight value  indicated the weight of abundance score in the IIS formula.Default value is 0.1}

\item{NS}{a number that represents the threshold value to select the difference from the results of IIS}

\item{r}{correlation coefficient of nodes.Default value is 0.9}

\item{p_adjust}{p value the have been adjusted.Default value is 0.05}

\item{MMpath}{the built-in dataset includes part of metabolic pathways and microbial pathways.}
}
\description{
A function that calculates and selects the key nodes,the formula of IIS is:
IIS= nDFS\emph{DFS + nDS}DS+nES\emph{ES+nAS}AS;
nDFS,nDS,nES,nAS indict the weight value of DFS,DS,ES and AS, and nDFS+nDS+nES+nAS=1;
DFS means Difference Score;
DS means Degree Score;
ES means Edge Score;
AS means Abundance Score.
}
\examples{
IIScore=IIS(Microbe_example,Meta_example,confounderData,groupInfo,MMpath=MMpathway3)
}
