% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\name{IPS}
\alias{IPS}
\title{IPS}
\usage{
IPS(TarnetPairs, IIS)
}
\arguments{
\item{TarnetPairs}{the results of microbial-metabolic pairs from TarNet function.}

\item{IIS}{the results from Nodescore function.}
}
\description{
A function used to calculate the importance pair score (IPS) of a network based on the return of the TarNet function.
}
\examples{
IIScore=IIS(Microbe_example,Meta_example,confounderData,groupInfo)
meta_mic2func=MMfunc(IIScore)
TarnetPairs=TarNet(IIScore,meta_mic2func)
IPSres=IPS(TarnetPairs,IIScore)
}
