% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_microbe}
\alias{KEGG_microbe}
\title{KEGG_microbe}
\format{
An object of class \code{list} of length 3.
}
\usage{
KEGG_microbe
}
\description{
KEGG_microbe
}
\examples{
  head(KEGG_microbe)
}
\keyword{datasets}
