% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta2pathway.R
\name{Meta2pathway}
\alias{Meta2pathway}
\title{Meta2pathway}
\usage{
Meta2pathway(MetaFrame, kg)
}
\arguments{
\item{MetaFrame}{Metabolic data in csv format that includes Metabolite abundance.}

\item{kg}{the built-in dataset includes metabolites and metabolic pathways.}
}
\value{
Metaandpath a data frame contained two parts which one is the pathway conversed and the other combines data from Metabolomics and pathway
}
\description{
A function that can transfer metabolite abundance into related pathway data.
using Meta2pathway you can get data containing metabolite abundance and pathway that metabolites belong to.
Note: Please use C number in KEGG database to replace the metabolite names.For example, C00002 means ATP or Adenosine 5'-triphosphate in KEGG. Try your best
to match the C number of metabolites for predicting pathway that metabolites belong to more accurately. if the metabolites do not have C numbers, keep the original names.
See data example "Meta2pathway_example" for details.
}
\examples{
Meta2pathway(Meta2pathway_example,KEGG)
}
