% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta2pathway_example}
\alias{Meta2pathway_example}
\title{Meta2pathway_example}
\format{
An object of class \code{data.frame} with 30 rows and 45 columns.
}
\usage{
Meta2pathway_example
}
\description{
Meta2pathway_example
}
\examples{
  head(Meta2pathway_example)
}
\keyword{datasets}
