% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_example}
\alias{Meta_example}
\title{Meta_example}
\format{
An object of class \code{data.frame} with 50 rows and 42 columns.
}
\usage{
Meta_example
}
\description{
Meta_example
}
\examples{
  head(Meta_example)
}
\keyword{datasets}
