% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicPathmatch.R
\name{MicPathmatch}
\alias{MicPathmatch}
\title{MicPathmatch}
\usage{
MicPathmatch(Microbelist, mic_pathways)
}
\arguments{
\item{Microbelist}{a list of microbe names in 16S or Metagenome data, and it does not contain the names of functional information in 16S or Metagenome data.}

\item{mic_pathways}{the built-in dataset includes complete microbial pathways from kegg}
}
\value{
Microberesult a list that contains two parts, one is a data frame that contains pathway and strains or sub-strains of microbes in the Microbelist
and the other is a data frame that only contains Microbelist names that need to predict and relative pathways
}
\description{
A function used to predict the functional pathway of microbes in the 16S or Metagenome data.
You can get two forms using this function to predict the possible next level strains and potental pathways of microbes you are interested in.
One is a form including predicted detail next level strains and pathways of your own microbes,
The other is a form including only your own microbes and corresponding pathways.
The format of microbe name is like this:
k__XXX; p__XXX; c__XXX; o__XXX; f__XXX; g__XXX; s__XXX;
k means kingdom;
p means phylum;
c means class;
o means order;
f means family;
g means genus;
s means species.
}
\examples{
Micrest=MicPathmatch(Microbe_path_example,mic_pathways=pathlist)
}
