% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microbe_example}
\alias{Microbe_example}
\title{Microbe_example}
\format{
An object of class \code{data.frame} with 50 rows and 27 columns.
}
\usage{
Microbe_example
}
\description{
Microbe_example
}
\examples{
  head(Microbe_example)
}
\keyword{datasets}
