% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microbe_path_example}
\alias{Microbe_path_example}
\title{Microbe_path_example}
\format{
An object of class \code{character} of length 30.
}
\usage{
Microbe_path_example
}
\description{
Microbe_path_example
}
\examples{
  head(Microbe_path_example)
}
\keyword{datasets}
