% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microbe_path_result}
\alias{Microbe_path_result}
\title{Microbe_path_result}
\format{
An object of class \code{list} of length 2.
}
\usage{
Microbe_path_result
}
\description{
Microbe_path_result
}
\examples{
  head(Microbe_path_result)
}
\keyword{datasets}
