% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMfunc.R
\name{MMfunc}
\alias{MMfunc}
\title{MMfunc}
\usage{
MMfunc(IIScore, NSpath = 2.5, NSmm = 2)
}
\arguments{
\item{IIScore}{the result using IIS() function.}

\item{NSpath}{a number that represents the threshold value to select the pathways from the results of complete nodescores}

\item{NSmm}{a number that represents the threshold value to select the Microbes/Metabolites from the results of complete nodescores}
}
\description{
A function that selects microbe and metabolites on the same pathway based on the data of Nodescore.The result will be saved as HTML file.
}
\examples{
IIScore=IIS(Microbe_example,Meta_example,confounderData,groupInfo)
meta_mic2func=MMfunc(IIScore)
}
