% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MMpathway}
\alias{MMpathway}
\title{MMpathway}
\format{
An object of class \code{data.frame} with 1029754 rows and 4 columns.
}
\usage{
MMpathway
}
\description{
MMpathway
}
\examples{
  head(MMpathway)
}
\keyword{datasets}
