% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MMpathway3}
\alias{MMpathway3}
\title{MMpathway3}
\format{
An object of class \code{data.frame} with 6149 rows and 4 columns.
}
\usage{
MMpathway3
}
\description{
MMpathway3
}
\examples{
  head(MMpathway3)
}
\keyword{datasets}
