% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathlist}
\alias{pathlist}
\title{pathlist}
\format{
An object of class \code{data.frame} with 1029758 rows and 10 columns.
}
\usage{
pathlist
}
\description{
pathlist
}
\examples{
  head(pathlist)
}
\keyword{datasets}
