% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarNet.R
\name{TarNet}
\alias{TarNet}
\title{TarNet}
\usage{
TarNet(IIScore, meta_mic2func, r = 0.1, p_adjust = 0.05)
}
\arguments{
\item{IIScore}{the result using IIS() function.}

\item{meta_mic2func}{the results of the complete nodescores getting from Nodescore function.}

\item{r}{correlation coefficient of nodes.Default value is 0.1.}

\item{p_adjust}{p value the have been adjusted.Default value is 0.05.}
}
\description{
A function used to plot a network that consists of microbes and metabolites in same pathway.This function offers a network of microbes and metabolites in same pathways.
}
\examples{
IIScore=IIS(Microbe_example,Meta_example,confounderData,groupInfo)
meta_mic2func=MMfunc(IIScore)
TarNet(IIScore,meta_mic2func)
}
